[the-graph-editor] Fire graphInitialised event
As the-graph-editor has null graph/nofloGraph properties until
the first time a graph is set, it's not possible for interested
parties to bind to events on the graph, or to know when they are
able to.

Add a graphInitialised event which those parties can listen for.
When it's fired, the listener knows the graph is "live" and can
be queried/bound to etc.